Use improved text from html
This is because the default plain text from the email is not of good
quality. There are a lot of connected words due to poor conversion.
We use beautifulsoup4 to parse the HTML and convert back to plain text,
adding a space in the middle to prevent words for sticking together.

Moreover, we condense multiple line feeds in the plain text into a
single space character, on the suspicion that this may cause the LLM to
unnecessarily switch context.